requirements.txt
# FastAPI and dependencies
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# Validation
pydantic>=2.6.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# Database
sqlalchemy>=2.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.0

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.0

# Utilities
python-dotenv>=1.0.0
httpx[socks]>=0.25.0
socksio>=1.0.0
aiofiles>=23.2.0
loguru>=0.7.0

# Image processing
Pillow>=10.2.0

# AI/ML
google-generativeai>=0.3.0
google-cloud-vision>=3.5.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0